+++ doc/source/conf.py
@@ -20,7 +20,8 @@
 import pathlib
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../'))
+
+sys.path.insert(0, os.path.abspath("../../"))
 
 # -- General configuration ------------------------------------------------
 
@@ -28,13 +29,12 @@ sys.path.insert(0, os.path.abspath('../../'))
 #
 # needs_sphinx = '1.0'
 
-autodoc_mock_imports = [
-]
+autodoc_mock_imports = []
 
 autodoc_default_options = {
-    'special-members': '__init__',
-    'undoc-members': True,
-    'show-inheritance': True,
+    "special-members": "__init__",
+    "undoc-members": True,
+    "show-inheritance": True,
 }
 
 autodoc_inherit_docstrings = False
@@ -43,40 +43,40 @@ autoclass_content = "class"
 
 # https://stackoverflow.com/questions/12772927/
 # Will only work for sphinx > 1.4 onward
-suppress_warnings = ['image.nonlocal_uri']
+suppress_warnings = ["image.nonlocal_uri"]
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.todo',
-    'sphinx.ext.coverage',
-    'sphinx.ext.mathjax',
-    'sphinx.ext.viewcode',
+    "sphinx.ext.autodoc",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.todo",
+    "sphinx.ext.coverage",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.viewcode",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # The encoding of source files.
 #
 # source_encoding = 'utf-8-sig'
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # General information about the project.
-project = 'RandomFileTree'
-copyright = '2019, Kilian Lieret'
-author = 'Kilian Lieret'
+project = "RandomFileTree"
+copyright = "2019, Kilian Lieret"
+author = "Kilian Lieret"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -134,7 +134,7 @@ exclude_patterns = []
 # show_authors = False
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 # A list of ignored prefixes for module index sorting.
 # modindex_common_prefix = []
@@ -151,20 +151,21 @@ todo_include_todos = True
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-#html_theme = 'alabaster'
+# html_theme = 'alabaster'
 
 # Add any paths that contain custom themes here, relative to this directory.
 # html_theme_path = ["_themes"]
 
 try:
     import importlib
-    theme = importlib.import_module('sphinx_rtd_theme')
-    html_theme = 'sphinx_rtd_theme'
+
+    theme = importlib.import_module("sphinx_rtd_theme")
+    html_theme = "sphinx_rtd_theme"
     html_theme_path = [theme.get_html_theme_path()]
 except ImportError:
     html_theme = "default"
 print("html_theme='{}'".format(html_theme))
-    
+
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
@@ -197,7 +198,7 @@ print("html_theme='{}'".format(html_theme))
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 # Add any extra paths that contain custom files (such as robots.txt or
 # .htaccess) here, relative to this directory. These files are copied
@@ -277,34 +278,36 @@ html_static_path = ['_static']
 # html_search_scorer = 'scorer.js'
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'RandomFileTreedoc'
+htmlhelp_basename = "RandomFileTreedoc"
 
 # -- Options for LaTeX output ---------------------------------------------
 
 latex_elements = {
-     # The paper size ('letterpaper' or 'a4paper').
-     #
-     # 'papersize': 'letterpaper',
-
-     # The font size ('10pt', '11pt' or '12pt').
-     #
-     # 'pointsize': '10pt',
-
-     # Additional stuff for the LaTeX preamble.
-     #
-     # 'preamble': '',
-
-     # Latex figure (float) alignment
-     #
-     # 'figure_align': 'htbp',
+    # The paper size ('letterpaper' or 'a4paper').
+    #
+    # 'papersize': 'letterpaper',
+    # The font size ('10pt', '11pt' or '12pt').
+    #
+    # 'pointsize': '10pt',
+    # Additional stuff for the LaTeX preamble.
+    #
+    # 'preamble': '',
+    # Latex figure (float) alignment
+    #
+    # 'figure_align': 'htbp',
 }
 
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'RandomFileTree.tex', 'RandomFileTree Documentation',
-     'Kilian Lieret', 'manual'),
+    (
+        master_doc,
+        "RandomFileTree.tex",
+        "RandomFileTree Documentation",
+        "Kilian Lieret",
+        "manual",
+    ),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -345,8 +348,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'randomfiletree', 'RandomFileTree Documentation',
-     [author], 1)
+    (master_doc, "randomfiletree", "RandomFileTree Documentation", [author], 1)
 ]
 
 # If true, show URL addresses after external links.
@@ -360,9 +362,15 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'RandomFileTree', 'RandomFileTree Documentation',
-     author, 'RandomFileTree', 'One line description of project.',
-     'Miscellaneous'),
+    (
+        master_doc,
+        "RandomFileTree",
+        "RandomFileTree Documentation",
+        author,
+        "RandomFileTree",
+        "One line description of project.",
+        "Miscellaneous",
+    ),
 ]
 
 # Documents to append as an appendix to all manuals.
@@ -436,7 +444,7 @@ epub_copyright = copyright
 # epub_post_files = []
 
 # A list of files that should not be packed into the epub file.
-epub_exclude_files = ['search.html']
+epub_exclude_files = ["search.html"]
 
 # The depth of the table of contents in toc.ncx.
 #
@@ -468,4 +476,4 @@ epub_exclude_files = ['search.html']
 
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/': None}
+intersphinx_mapping = {"https://docs.python.org/": None}