+++ randomfiletree/__init__.py
@@ -1,2 +1,7 @@
-from randomfiletree.core import *
-from randomfiletree.cli import *
+from randomfiletree.core import (  # noqa F401
+    iterative_tree,
+    iterative_gaussian_tree,
+    choose_random_elements,
+    sample_random_elements,
+    random_string,
+)